Moving sanity checks to interpro.py
