+++ pronto/views/api/__init__.py
@@ -2,7 +2,7 @@ from cx_Oracle import DatabaseError
 from flask import jsonify, request
 
 from pronto import app, db, executor, get_user
-from . import (annotation, database, entry, interpro, protein, sanitycheck,
+from . import (annotation, database, entry, interpro, protein,
                signature, signatures, uniprot)
